refactor(nfqws2): extract engine file/version/update/reload into internal/nfqws2ctl
The nfqws2 file-management + package version/update/SIGHUP-reload surface moves out
of internal/app (was app/nfqws2.go + nfqws2_{linux,other}.go, ~520 lines) into a
dedicated internal/nfqws2ctl package with a Manager type. App holds the Manager +
11 thin delegators (+ Nfqws2File/Nfqws2VersionInfo type aliases) so the
internal/server handlers are unchanged. Functions moved verbatim; the package gets
its own opkgBin/lastLines copies to decouple, and app keeps opkgBin (still used by
the AWG2 engine install). Start/Stop stay in services.go (service-control).

Verified: go build (linux) + go vet (host) + go test pass; live router smoke —
file list / version / file read / reload all identical.

First of the per-service package split (nfqws2 -> monitor -> blobs -> proxy -> awgroute).
